src/requirements.txt
asgiref==3.8.1
Django==4.2.17
django-cors-headers==4.4.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework-jwt==1.11.0
psycopg2-binary==2.9.9
PyJWT==1.7.1
python-multipart==0.0.20
sqlparse==0.5.1
typing_extensions==4.12.2